Refactor MetadataFile class
