+++ src/nodeserver/wrapper/metadata/base_metadata.py
@@ -5,4 +5,3 @@ class BaseMetadata(BaseModel):
     capitalized_name: str
     description: str = ""
 
-
